Dynamically inserted pagination and dynamic search functionality
The application takes a list of students, organizes them by groups of
10 student elements across as many pages as necessary, all of which are
accessible and marked active by user interaction while also allowing
for dynamic search of the student elements by the user. The searched
elements are paginated. A blank search input element upon button click
results in a student element list reset (all student elements are shown
and paginated).
————————————

Pagination works properly upon initialization but not after a blank
search is run. Requires small edit to function name so that it can be
called within the search function to properly paginate shown student
list elements.

Search works and displays elements. Undetermined if elements will
become paginated if more than 10 elements are returned, but it is
likely that pagination function will have to be called after search
results are shown based on number of search results.

Search still needs to be updated to work dynamically when the user lets
up a key on the keyboard.

Animations still need to be added when proper search results are shown.

Message still needs to be created and added if there are no search
results given the user’s input and click of search button.